inverted pixel clock for capture support
support setting inverted pixel clock for capture.
This is needed as a workaround for a pixel clock glitch in
the "new" (spring 2018) TRIK video modules.
The modules seem to have the glitch during the start of line
code but only for normal clock polarity.

The clock can be set normal or inverted for both channels at once only.

Set it via the v4l2 S_CTRL ioctl (v4l2-ctl program).

The value is supposed to be set from a control script based on
whether the script runs on a "new" or "old" controller.