Backport upstream commit for dacinci i2c
Author: Alexander Sverdlin <alexander.sverdlin@nokia.com>
Date:   Tue Jun 9 13:10:10 2015 +0200

i2c: davinci: Avoid sending to own address

Sending a message to own address locks the controller up in very bizarre state,
it behaves as slave even if MDR register clearly states master. The controller
remains in this state until reset. To avoid unnecessary timeouts simply avoid
sending to own address. The controller cannot do this any way. Also, do not
enable AAS IRQ, as the slave mode is not supported by the driver and the only
possibility to trigger this IRQ is to send to own address.

Signed-off-by: Alexander Sverdlin <alexander.sverdlin@nokia.com>
Signed-off-by: Wolfram Sang <wsa@the-dreams.de>